requirements.txt
langchain-chroma
langchain-core
langchain-google-genai
langchain-text-splitters
python-dotenv